requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
gevent==24.2.1
greenlet==3.0.3
idna==3.8
numpy==2.1.1
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
setuptools==74.1.2
six==1.16.0
TA-Lib==0.4.32
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
v20==3.0.25.0
websocket-client==1.8.0
zope.event==5.0
zope.interface==7.0.3
